공 트래킹 및 바운드완료
