--- architecture/Web_UI.drawio
@@ -1,1 +1,0 @@
-<mxfile host="app.diagrams.net" modified="2022-01-10T07:12:31.366Z" agent="5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/96.0.4664.110 Safari/537.36" etag="s2tAmQr51CMTCO7R3s4h" version="16.2.2" type="device"><diagram id="73ZQPh2pG8uffAKDGe0n" name="Page-1">7V3rk6I6Fv9rqNr9gEUeQPjoo53Z2nuruu69O49PW7SisoPiAj3dvX/9JpAoJFFRAemxnalqCBjD+Z2T88jJwUDj9eunxN+ufo/nQWRAa/5qoIkBoU0sTP+wlreiBQEAi5ZlEs6LNrBv+DP8X8AbLd76HM6DtHJjFsdRFm6rjbN4swlmWaXNT5L4pXrbIo6qv7r1l4HS8OfMj9TWr+E8WxWtBLr79s9BuFyJXwaOV1xZ++Jm/iTpyp/HL6Um9GCgcRLHWXG0fh0HEaOeoEvxvemBq7uBJcEmq/OFFzyN//gX+J466/Vn8zOYPz5/MXkvP/3omT8wH2z2JiiQxM+becA6sQw0elmFWfDn1p+xqy8UdNq2ytYRPQP0UB2U+IUgyYLXUhMf5KcgXgdZ8kZv4VdNgAXJBNMQTtOXPQLY4fesStSHhDf6HPXlrvc9YegBp80ZdII9pJOLJDJ51q3JhDRkciL6s6NFTB+0TC/nv8+xuGCmudwP6Q2YbF/3F+nRMv/7MDRG0BhN8gNijOgDWMbDyBg6xnBsPIwNMjSG2HiYsAN6Q3GVQMMj7KrnGiMshkKfrBgN7/vWQDpOFUf35jDi/nG7xOvYujmRnKt5HWI9r08Mb2iQnK29kUGlmjG9ZdAnYS0PBrHrsjIlblaFIs2S+EcwjqM4oS2beBOwcYVRJDX5Ubjc0NMZhSyg7SMGVUj14pBfWIfzOfsZLfZV7mhGKQC3Otu5WOEAAHUc0BYDuK0xAJ3XgOGNc9xdNp0x3Cf04G/5Nc/w8K7p7/fHC4ri08yYAHfJCuQgK6Rbf3MNK6h6b4d20fWvjrasHW0Va+IN7A7B9tqU+4ps351ogxqC3ekcD2o4SFGYk7IguXASYQt+EXRg1QjCGhWoM4JAa+Sp4RfdjjzEvjV5dP5QX8hjw5tzTw0/43bkcW5OHrvH5HEsNSbTMXmud8CAVg//Y3J3epfi60qTp2Z26NSkBte7V3p4H79+wAuxo+rGbuG93mXSwyuMZ8zDgwTfncukwm1bt5bmw07TtXCPDEL/e/kBzoPARZDsPnGXnGXsqLiTDnEXHbci5sThUVJ2MLlLuKsxUVuzUGZ3KeZQ5y6fqbQPLACJyDeb1S1jSNVXEQ73mKSzRsSY4ewlHnAArRJ7+MmMr0sj1BhyuCqoDtJ4G5YOOrct6HrkytsEVa0VrIkDdepswB558jJ1bM1aWMfU0TjyhxcxrpPGRijoSBRENeOMqDUKanz9WuuAvSOlqwm6dUvKj7hAcyaGMhNDVVI6dSTgR1igRXTJjdFFOjdRIn+wmQ9ZYh+jY+SnaTirgkHpkbx9Y1QaWNTt5Q3f8wZku6Jh8soJWZy9lc8egySkD8QgKhpfw6zo0fMwP+cdOvx03x07eSudyJ0dBC6Nn5NZcIQ6Iv/RT5ZBdkxGivuCeSW1UWWDE16CaEuCyM/Cn0FlsDro+S88xmEuiILLPLtq62JPYp/iwfnX9hyk9kTNZqfSlSeSQERXBW2UrijD+G+l27bshlRh1x1trkidqrGmcZKDObsBysB2hd0ci5zNbxV5AFV5cIr+z5WHu2NhAJA7cF1v93ElNkQDr/SxL+VvgEilY6cmf1/AvY+xHf0zCFcP9peZtXl+9MYTUicd+BTzVrXUjvlMxn1Qmo0RPMV9pYnXq068rnuZJDTHy7AmL+N+8bIcI8TuhdMxPtFPy7xaI15xJq8KZqswmu2Csxntcp7qGa/IUSrrUs2tBKadbplFp5WbSUyfsIQtHoIs4h20xc6XH/IcdOawT/kxmbDT/QLVhN3Gc9nHxtA1PMC/PvRYC0v9ez9p647GeWgtJVuLsi5hopk0zHG+tpTDRAjbVHBviw0K2pqQWXt5mFq0dSvIjSgAUFEA0D4+/1csnJ17yZUHdrq1r3FNmwT1Ss/Yco6vPEPU1TOOlEBqy1HGs/3DegMWts+hccn3Q76esGf/YgSN6rwGYil15MOtKR9W/0TjmKnQU9Fw8KWiIW2qdOTVoJZEQ/xOXdHwQAeiIZ69eUuB9/KUyC2nLIWuzTeApLBcp9sO9agczh641n6jdrgjrHT7Hu03Be5O99Ho4W7Pgx94ngiUFtoGO+i6wKlk2GG348Apqqm9xMzWE/UFkOT4o0uD/2pVAFkTth0bbW9zO3UqkUFAHiIY5+v7edxg6InIgJXPXcUBEIGFIvNpaIxslhvQ0ygB8EgVNrMPiuZwOKiBQEEBZR4xGDG07k3TqIhrClN0rWsOh4YaCAA+sKDQrvYEszFcJpVwnItqHtDzSCUxnR2Mq5HA4RkbPm8uxT2Q4cNpPk3LMLxDGXarC5E9sBZ1iT+/eLjvtLXXrzgeIIBRZf+xK0wE5GhcbdvPkyo+ANmKbCh4AYg0z4khHxyZ/AXqxLoSr7cRv+go9n1haO+EEBzk97IXdNS7OS0Y9nsSDHjpEryil5Uc77YEA54I6ykiK33harn4ChYZAtGX6et/xqPV5NPs27fFeeXsusmQBuh0jqU2QRrIS+SXqFItlWrEXbqmEnTcvlGpRiJfuvK37DBc51UkyzSR7bcs3pZaf/OfgugxTsMsjNnVpzjL4jW9IWIXRv7sxzJHQFiJBkSL/ENvyX9smG6LapdsPvXFySJ8ZZiN+HgmqyxjZTKHjBBwOptv4CCcxZtFSLFNBjP6i3A69zOf/mHtae52xNmTH0Umm5KmiNFwEQbR3ASQDLabZa44Mp+P22vI3oTSLgLsqfsW6QA08IvGxuGvUXxBwL9+XbLqp4OUQp0O0m2cZOkgCzabMK0yRT7LBsnDz4BNtkVb2RmgMBfq4ZDncJJ/9GyneBdNzGsYDapazAbWQN384XkqaqKtcdDODClzqtbMXpRyZwHyWjP230Fu2C6LW4gskkCta8Mo+YiyC9lcXFfPMzW29t21pO/2F4rAu+iiEvxRGQ23JeWwhpR/qOaGVLMs5y5QLTNHo5mdthRzna24ly4bVnZb1PR99+7ud6O0A+ka37du/oqePv3SE3ZTu3+glChi1kwhPtfVlX+HHHd0TQzQkfvb8XN1G64bW4uktnexJMESm8eVKtujsdh9XCxbjNkeZK8otz1m5baL27yhWLssb1jerX0glibd05UO08FVOE1bt+W1tbUOPdztJTbXAeg+1jwo8u7AtSXwO1340IOvW+i6Tt/tdRcmpKK9qJtDTmiwy70a/eN5vVJXynRuXppwCXA1pGUqJQ8ai8262hG3q4GaX4Fo2qI6yXoI/KKsZxJpY7+5G2PDvGcCCxDtoNvlvuvLydUvuF4oRyIyApBIwcpfK8IrUhU7vyTVqSYB962lb0nJJiIDKKvg29tfYuLukN2Oc9b9GGWKNa4xyHZ5BZ3YY0Jr9Mz3Kmoh5uZ8j70rsCs/2Sfx1sUTG0slOwTMnQiwBvHb+1SozuLujyCbrTgldPH83Qv92A1zP13t6HYObAZEtsX+iRBzKfasYHjmakD8nLGihOPduxEtHYPIixvxYhHOgsEs3syCbZYOfobzIP73NmJGXBMsgYDsr2gYQmdny+Zwc+xQZ7H3gx06YgfkoRuzQ42Iywc7dMYOmtK43bJDjcxjbe3XIsaR+Zn/lN+ncIKMaFI8zQ6EKFhkxzQ4NSFm4Wb5F4Mz75A3/JZ/b+II1swpYo8Me5K3JBnFnA7VD3OUAj/NXoI0awY/ZXLXrBfuTLmmq9Pq8asRqvrA7xB+yFFn447xq1HI4AO/g/hpXv3QLX5YjZ4ABUDaT7hND7kl5WrsckaFWm+b+WZTfx1GjAZ/+at47e+xGXGlOMH7tj84FaCikqn+nef/mopvSZMjVPMc23q/nR4bNZihCtedYIMlwXHVlyl0i40alkD3ig2UsNEEjbrF5vrN4PfwoguAgbKkDnWp/W296kIPnu51BWqk7hpSNkA7ucSUxo7Wk22Xhd084XRvKTiSQXJjCsqvk+0FCTXO5LH1nhuTEDm4fyQ8/Oq4Bl7APBS1N+nc29/9+DWKN+ls/PbWULDOSWtKIVa346NWlB/riWs/iBtCyZEy5zWaz9Wh5LQEkq16Ym2BBN8NSESqKwRujtL1C8x1UQLvBiWMPGm72c1hUt0zhWzvbheKdXIXytNzGm6CNDVncRTRXun4TKpQqVkzZRbrdBsG5mzlJ5mJShtTSiwBSFMsIb3kFOjeEtfxptEz9pLdEU/kK0cmbJkfHKXiALw5P1y/T+PXm8qJ5JaZmiBotzN5e8XBStlaRRE/vimmmvPHwztF8t+7reynfTdct5lcdndFwcpeyH2kcilFwbDmxaMdJ3LZjRc5lzVTs0XPjS4rIx2d7Hqy6QEQefv/xfttPLlAbFdFwk5tEVW+IGqet7pHwvlw9tUpzJZ45NZOpNOdr/9+wmaygYgtrCks0xBO9DSJGen2sse8ld/jecDu+D8=</diagram></mxfile>
