--- architecture/analysis application map.drawio
@@ -1,1 +1,0 @@
-<mxGraphModel><root><mxCell id="0"/><mxCell id="1" parent="0"/></root></mxGraphModel>
