+++ document/architecture/analysis application map.drawio
@@ -1,0 +1,1 @@
+<mxGraphModel><root><mxCell id="0"/><mxCell id="1" parent="0"/></root></mxGraphModel>
