--- tennis_tracking/TrackPlayers/trackplayers.py
@@ -1,235 +1,0 @@
-from collections import OrderedDict
-
-import cv2
-import numpy as np
-from scipy.spatial import distance as dist
-
-
-def get_output_layers(net):
-    layer_names = net.getLayerNames()
-    output_layers = [layer_names[i[0] - 1] for i in net.getUnconnectedOutLayers()]
-    return output_layers
-
-
-def binary(x):
-	# transform an image in black and white
-    if x == 0:
-        return 0
-    else:
-        return 255
-
-
-def remove_ball_boy(detected_person_img, lower_col, upper_col):
-	# compute pixel percentage of a range of color (lower_col, upper_col)
-	# in each box to detect ball boys/girls
-    mask = cv2.inRange(detected_person_img, lower_col, upper_col)
-    img = cv2.bitwise_and(detected_person_img, detected_person_img,\
-    	                  mask=mask)
-    func = np.vectorize(binary)
-    img = func(img).astype(np.uint8)
-
-    n_pix = img.shape[0] * img.shape[1]
-    n_pix_bb = sum(img.flatten()) / img.shape[2] / 255
-
-    return n_pix_bb / n_pix
-
-
-def predict_players(outs, LABELS, img, confidence_threshold=0.8):
-    class_ids = []
-    confidences = []
-    boxes = []
-    conf_threshold = 0.8
-    nms_threshold = 0.00000000001
-    Width = img.shape[1]
-    Height = img.shape[0]
-
-    predicted_players = []
-
-    for out in outs:
-        for detection in out:
-            scores = detection[5:]
-            class_id = np.argmax(scores)
-            confidence = scores[class_id]
-
-            if confidence > conf_threshold and LABELS[class_id] == 'person':
-                center_x = int(detection[0] * Width)
-                center_y = int(detection[1] * Height)
-                w = int(detection[2] * Width)
-                h = int(detection[3] * Height)
-                x = int(center_x - w / 2)
-                y = int(center_y - h / 2)
-
-                taux = remove_ball_boy(img[y:y + h, x:x + w],\
-                					  (27, 5, 40), (47, 40, 168))
-
-                if taux < 0.01:
-                    class_ids.append(class_id)
-                    confidences.append(float(confidence))
-                    boxes.append([x, y, w, h])
-
-    # apply non-maxima suppression to suppress weak, overlapping
-    # bounding boxes
-    idxs = cv2.dnn.NMSBoxes(boxes, confidences, conf_threshold,
-                            nms_threshold)
-
-    # ensure at least one detection exists
-    if len(idxs) > 0:
-        # loop over the indexes we are keeping
-        for i in idxs.flatten():
-            # extract the bounding box coordinates
-            (x, y) = (boxes[i][0], boxes[i][1])
-            (w, h) = (boxes[i][2], boxes[i][3])
-
-            predicted_players.append((x, y, w, h))
-
-    # sometimes, two identical frames remain
-    predicted_players = list(set(predicted_players))
-
-    return predicted_players
-
-
-def update_boxes(box):
-    x = box[0]
-    y = box[1]
-    w = box[2]
-    h = box[3]
-    return x, y, x + w, y + h
-
-
-
-class CentroidTracker():
-    def __init__(self, maxDisappeared=300):
-        # initialize the next unique object ID along with two ordered
-        # dictionaries used to keep track of mapping a given object
-        # ID to its centroid and number of consecutive frames it has
-        # been marked as "disappeared", respectively
-        self.nextObjectID = 0
-        self.objects = OrderedDict()
-        self.disappeared = OrderedDict()
-        # store the number of maximum consecutive frames a given
-        # object is allowed to be marked as "disappeared" until we
-        # need to deregister the object from tracking
-        self.maxDisappeared = maxDisappeared
-
-    def register(self, centroid):
-        # when registering an object we use the next available object
-        # ID to store the centroid
-        self.objects[self.nextObjectID] = centroid
-        self.disappeared[self.nextObjectID] = 0
-        self.nextObjectID += 1
-
-    def deregister(self, objectID):
-        # to deregister an object ID we delete the object ID from
-        # both of our respective dictionaries
-        del self.objects[objectID]
-        del self.disappeared[objectID]
-
-    def update(self, rects):
-        # check to see if the list of input bounding box rectangles
-        # is empty
-        if len(rects) == 0:
-            # loop over any existing tracked objects and mark them
-            # as disappeared
-            for objectID in list(self.disappeared.keys()):
-                self.disappeared[objectID] += 1
-                # if we have reached a maximum number of consecutive
-                # frames where a given object has been marked as
-                # missing, deregister it
-                if self.disappeared[objectID] > self.maxDisappeared:
-                    self.deregister(objectID)
-            # return early as there are no centroids or tracking info
-            # to update
-            return self.objects
-
-        # initialize an array of input centroids for the current frame
-        inputCentroids = np.zeros((len(rects), 2), dtype="int")
-        # loop over the bounding box rectangles
-        for (i, (startX, startY, endX, endY)) in enumerate(rects):
-            # use the bounding box coordinates to derive the centroid
-            cX = int((startX + endX) / 2.0)
-            cY = int(endY)
-            inputCentroids[i] = (cX, cY)
-
-        # if we are currently not tracking any objects take the input
-        # centroids and register each of them
-        if len(self.objects) == 0:
-            for i in range(0, len(inputCentroids)):
-                self.register(inputCentroids[i])
-
-        # otherwise, are are currently tracking objects so we need to
-        # try to match the input centroids to existing object
-        # centroids
-        else:
-            # grab the set of object IDs and corresponding centroids
-            objectIDs = list(self.objects.keys())
-            objectCentroids = list(self.objects.values())
-            # compute the distance between each pair of object
-            # centroids and input centroids, respectively -- our
-            # goal will be to match an input centroid to an existing
-            # object centroid
-            D = dist.cdist(np.array(objectCentroids), inputCentroids)
-            # in order to perform this matching we must (1) find the
-            # smallest value in each row and then (2) sort the row
-            # indexes based on their minimum values so that the row
-            # with the smallest value is at the *front* of the index
-            # list
-            rows = D.min(axis=1).argsort()
-            # next, we perform a similar process on the columns by
-            # finding the smallest value in each column and then
-            # sorting using the previously computed row index list
-            cols = D.argmin(axis=1)[rows]
-
-            # in order to determine if we need to update, register,
-            # or deregister an object we need to keep track of which
-            # of the rows and column indexes we have already examined
-            usedRows = set()
-            usedCols = set()
-            # loop over the combination of the (row, column) index
-            # tuples
-            for (row, col) in zip(rows, cols):
-                # if we have already examined either the row or
-                # column value before, ignore it
-                # val
-                if row in usedRows or col in usedCols:
-                    continue
-                # otherwise, grab the object ID for the current row,
-                # set its new centroid, and reset the disappeared
-                # counter
-                objectID = objectIDs[row]
-                self.objects[objectID] = inputCentroids[col]
-                self.disappeared[objectID] = 0
-                # indicate that we have examined each of the row and
-                # column indexes, respectively
-                usedRows.add(row)
-                usedCols.add(col)
-
-            # compute both the row and column index we have NOT yet
-            # examined
-            unusedRows = set(range(0, D.shape[0])).difference(usedRows)
-            unusedCols = set(range(0, D.shape[1])).difference(usedCols)
-
-            # in the event that the number of object centroids is
-            # equal or greater than the number of input centroids
-            # we need to check and see if some of these objects have
-            # potentially disappeared
-            if D.shape[0] >= D.shape[1]:
-                # loop over the unused row indexes
-                for row in unusedRows:
-                    # grab the object ID for the corresponding row
-                    # index and increment the disappeared counter
-                    objectID = objectIDs[row]
-                    self.disappeared[objectID] += 1
-                    # check to see if the number of consecutive
-                    # frames the object has been marked "disappeared"
-                    # for warrants deregistering the object
-                    if self.disappeared[objectID] > self.maxDisappeared:
-                        self.deregister(objectID)
-
-            # otherwise, if the number of input centroids is greater
-            # than the number of existing object centroids we need to
-            # register each new input centroid as a trackable object
-            else:
-                for col in unusedCols:
-                    self.register(inputCentroids[col])
-        # return the set of trackable objects
-        return self.objects
